Don't use the package parser
I think at the moment we're using a that will consider anything that can be
decoded as an *Unstructured as both package metadata and a package (payload)
object, so I think we're actually just outputting what we consider to be a big
stream of package metadata files. Avoiding the parser would make the code a bit
easier to follow (it acts like it's loading objects, but it doesn't).

This commit also adds some tests, and ensures that package output is sent to
stdout while all other output (e.g. errors) are sent to stderr.

Signed-off-by: Nic Cope <negz@rk0n.org>